SCRUM-1046 added login to UI
