+++ src/main/cliapp/src/App.js
@@ -1,22 +1,13 @@
-import React, { useState, useEffect, useRef } from 'react';
-import { Route, useHistory, Switch } from 'react-router-dom';
-import { CSSTransition } from 'react-transition-group';
+import React from 'react';
+import { useHistory } from 'react-router-dom';
+import { Security } from '@okta/okta-react';
+import { OktaAuth } from '@okta/okta-auth-js';
+import { oktaAuthConfig } from './oktaAuthConfig';
 
-import { Security, SecureRoute, LoginCallback, useOktaAuth } from '@okta/okta-react';
-import { OktaAuth, toRelativeUrl } from '@okta/okta-auth-js';
-import { oktaAuthConfig, oktaSignInConfig } from './oktaAuthConfig';
-
-import classNames from 'classnames';
-
-
-import { Login } from './Login';
+import { SiteLayout } from './containers/layout'
 
 import routes from './routes';
 
-
-
-
-import PrimeReact from 'primereact/api';
 
 import 'primereact/resources/themes/saga-blue/theme.css';
 import 'primereact/resources/primereact.min.css';
@@ -30,33 +21,28 @@ import './App.scss';
 
 const App = () => {
 
-    const oktaAuth = new OktaAuth(oktaAuthConfig);
+  const oktaAuth = new OktaAuth(oktaAuthConfig);
 
-    //const { oktaAuth, authState } = useOktaAuth();
+  const history = useHistory();
 
-    const history = useHistory();
+  const customAuthHandler = () => {
+    history.push('/login');
+  };
 
-    const customAuthHandler = () => {
-      history.push('/login');
-    };
+  const restoreOriginalUri = async (_oktaAuth, originalUri) => {
+    console.log(originalUri);
+    console.log(window.location);
+    //history.replace(toRelativeUrl(originalUri, window.location.origin));
+  };
 
-    const restoreOriginalUri = async (_oktaAuth, originalUri) => {
-      console.log(originalUri);
-      console.log(window.location);
-      //    history.replace(toRelativeUrl(originalUri, window.location.origin));
-    };
 
-    return (
-      <>
-        { console.log(oktaAuth.isAuthenticated()) }
-        <Security oktaAuth={oktaAuth} onAuthRequired={customAuthHandler} restoreOriginalUri={restoreOriginalUri}> 
-            { routes }
-            <Route path='/login' render={() => <Login config={oktaSignInConfig} />} />
-        </Security>
-      </>
-
-    );
-
+  return (
+    <Security oktaAuth={oktaAuth} onAuthRequired={customAuthHandler} restoreOriginalUri={restoreOriginalUri}>
+      <SiteLayout>
+        {routes}
+      </SiteLayout>
+    </Security>
+  );
 };
 
 export default App;
