+++ src/main/cliapp/src/routes.js
@@ -1,11 +1,10 @@
-import React, { useState, useEffect, useRef } from 'react';
-import { Route, useHistory, Switch } from 'react-router-dom';
-import { CSSTransition } from 'react-transition-group';
-import { Security, SecureRoute, LoginCallback } from '@okta/okta-react';
-
-import { Layout, Layout2 } from './containers/layout';
+import React from 'react';
+import { Route } from 'react-router-dom';
+import { SecureRoute } from '@okta/okta-react';
+import { oktaSignInConfig } from './oktaAuthConfig';
 
 import { DashboardPage } from './containers/dashboardPage';
+import { Login } from './Login'
 
 import { DataLoadsPage } from './containers/dataLoadsPage/';
 import { DiseaseAnnotationsPage } from './containers/diseaseAnnotationsPage';
@@ -34,32 +33,29 @@ import { ZECOOntologyComponent } from './containers/ontologies/ZECOOntologyCompo
 
 
 export default (
-  <Layout>
-    <Switch>
-      <SecureRoute path="/" exact component={DashboardPage} />
-      <SecureRoute path="/dataloads" component={DataLoadsPage} />
-      <SecureRoute path="/diseaseAnnotations" component={DiseaseAnnotationsPage} />
-      <SecureRoute path="/genes" component={GenesPage} />
-      <SecureRoute path="/alleles" component={AllelesPage} />
-      <SecureRoute path="/molecules" component={MoleculesPage} />
-      <SecureRoute path="/vocabterms" component={ControlledVocabularyPage} />
-      <SecureRoute path="/ontology/chebi" component={CHEBIOntologyComponent} />
-      <SecureRoute path="/ontology/do" component={DiseaseOntologyComponent} />
-      <SecureRoute path="/ontology/eco" component={ECOOntologyComponent} />
-      <SecureRoute path="/ontology/go" component={GOOntologyComponent} />
-      <SecureRoute path="/ontology/so" component={SOOntologyComponent} />
-      <SecureRoute path="/ontology/ma" component={MAOntologyComponent} />
-      <SecureRoute path="/ontology/zfa" component={ZFAOntologyComponent} />
-      <SecureRoute path="/ontology/mp" component={MPOntologyComponent} />
-      <SecureRoute path="/ontology/dao" component={DAOOntologyComponent} />
-      <SecureRoute path="/ontology/emapa" component={EMAPAOntologyComponent} />
-      <SecureRoute path="/ontology/wbbt" component={WBbtOntologyComponent} />
-      <SecureRoute path="/ontology/xco" component={XCOOntologyComponent} />
-      <SecureRoute path="/ontology/zeco" component={ZECOOntologyComponent} />
-      <SecureRoute path="/fmspage" component={FMSComponent} />
-      <SecureRoute path="/agms" component={AffectedGenomicModelPage} />
-    </Switch>
-  </Layout>
-
-
+  <>
+    <Route path='/login' render={() => <Login config={oktaSignInConfig} />} />
+    <SecureRoute path="/" exact component={DashboardPage} />
+    <SecureRoute path="/dataloads" component={DataLoadsPage} />
+    <SecureRoute path="/diseaseAnnotations" component={DiseaseAnnotationsPage} />
+    <SecureRoute path="/genes" component={GenesPage} />
+    <SecureRoute path="/alleles" component={AllelesPage} />
+    <SecureRoute path="/molecules" component={MoleculesPage} />
+    <SecureRoute path="/vocabterms" component={ControlledVocabularyPage} />
+    <SecureRoute path="/ontology/chebi" component={CHEBIOntologyComponent} />
+    <SecureRoute path="/ontology/do" component={DiseaseOntologyComponent} />
+    <SecureRoute path="/ontology/eco" component={ECOOntologyComponent} />
+    <SecureRoute path="/ontology/go" component={GOOntologyComponent} />
+    <SecureRoute path="/ontology/so" component={SOOntologyComponent} />
+    <SecureRoute path="/ontology/ma" component={MAOntologyComponent} />
+    <SecureRoute path="/ontology/zfa" component={ZFAOntologyComponent} />
+    <SecureRoute path="/ontology/mp" component={MPOntologyComponent} />
+    <SecureRoute path="/ontology/dao" component={DAOOntologyComponent} />
+    <SecureRoute path="/ontology/emapa" component={EMAPAOntologyComponent} />
+    <SecureRoute path="/ontology/wbbt" component={WBbtOntologyComponent} />
+    <SecureRoute path="/ontology/xco" component={XCOOntologyComponent} />
+    <SecureRoute path="/ontology/zeco" component={ZECOOntologyComponent} />
+    <SecureRoute path="/fmspage" component={FMSComponent} />
+    <SecureRoute path="/agms" component={AffectedGenomicModelPage} />
+  </>
 )
